requirements.txt
djangorestframework_datatables>=0.5.0
Django>=2.2.7
djangorestframework>=3.10.3
